+++ frontend/src/main/java/org/abs_models/backend/java/lib/runtime/AbstractAsyncCallRT.java
@@ -5,11 +5,9 @@
 package org.abs_models.backend.java.lib.runtime;
 
 import org.abs_models.backend.java.JavaBackendException;
-import org.abs_models.backend.java.lib.types.ABSBool;
 import org.abs_models.backend.java.lib.types.ABSDataType;
 import org.abs_models.backend.java.lib.types.ABSRational;
 import org.abs_models.backend.java.lib.types.ABSRef;
-
 import org.apfloat.Aprational;
 
 public abstract class AbstractAsyncCallRT<T extends ABSRef> extends AbstractAsyncCall<T> implements AsyncCallRTAttributes {
@@ -32,7 +30,7 @@ public abstract class AbstractAsyncCallRT<T extends ABSRef> extends AbstractAsyn
      */
     private boolean critical;
 
-    public AbstractAsyncCallRT(ABSObject source, T target, ABSDataType dl, ABSDataType co, ABSBool cr) {
+    public AbstractAsyncCallRT(ABSObject source, T target, ABSDataType dl, ABSDataType co, boolean cr) {
         super(source, target);
         deadline_t = convertFromDuration(dl);
         if (deadline_t.signum() > 0) {
@@ -40,7 +38,7 @@ public abstract class AbstractAsyncCallRT<T extends ABSRef> extends AbstractAsyn
             deadline_t = deadline_t.add(ABSRuntime.getRuntime().getClock());
         }
         cost = convertFromDuration(co);
-        critical = cr.toBoolean();
+        critical = cr;
     }
 
     @Override