+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The Java backend used subclasses of `ABSValue` to wrap instances of
 String, boolean etc., unwrapping them for all computations.  Use the
 wrapped values directly instead in generated code.
 - Instead of `ABSString`, use `java.lang.String`
+- Instead of `ABSBool`, use `java.lang.Boolean`
 
 ### Removed